+++ dio/lib/src/headers.dart
@@ -15,6 +15,7 @@ class Headers {
   // Header field value
   static const jsonContentType = 'application/json; charset=utf-8';
   static const formUrlEncodedContentType = 'application/x-www-form-urlencoded';
+  static const textPlainContentType = 'text/plain';
 
   static final jsonMimeType = MediaType.parse(jsonContentType);