+++ dio/lib/src/transformer.dart
@@ -34,11 +34,11 @@ abstract class Transformer {
 
   /// Deep encode the [Map<String, dynamic>] to percent-encoding.
   /// It is mostly used with  the "application/x-www-form-urlencoded" content-type.
-  static String urlEncodeMap(Map map,[CollectionFormat collectionFormat = CollectionFormat.multiCompatible]) {
+  static String urlEncodeMap(Map map,[ListFormat collectionFormat = ListFormat.multi]) {
     return encodeMap(map, (key, value) {
       if (value == null) return key;
       return '$key=${Uri.encodeQueryComponent(value.toString())}';
-    },collectionFormat: collectionFormat);
+    },listFormat: collectionFormat);
   }
 }