+++ dio/test/request_test.dart
@@ -149,5 +149,6 @@ void main() {
       assert(response.data is List);
       expect(response.data[0], 1);
     });
+
   });
 }